refactor: 네일 세트 상세, 리스트를 위젯에서 페이지로 전환 [SN-50]
